+++ src/main/java/com/book/store/Repository/CustomerRepository.java
@@ -13,9 +13,9 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
     @Lock(LockModeType.PESSIMISTIC_WRITE)
-    Optional<Customer> findById(Integer id);
+    Optional<Customer> findById(Long id);
 
     Optional<Customer> findByEmail(String email);
     Customer save(Customer customer);
